ai-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.33.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
h11==0.16.0
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
